requirements.txt
argon2-cffi==23.1.0
cryptography==43.0.1
PyNaCl==1.5.0
click==8.1.7
